DetAny3D/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
appdirs==1.4.4
aria2p==0.12.0
asciimatics==1.15.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.13.3
black==24.8.0
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cityscapesScripts==2.2.3
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.11
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
dill==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.8
fastcore==1.7.1
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.14.0
fire==0.7.0
flake8==7.1.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
ftfy==6.2.3
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==5.2.0
geos==0.2.3
glob2==0.7
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
gradio_image_prompter==0.1.0
-e git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44#egg=groundingdino
grpcio==1.67.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.2
humanfriendly==10.0
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
iopath==0.1.9
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keras==2.13.1
kitti360Scripts @ git+https://github.com/autonomousvision/kitti360Scripts.git@40ee015de8277b2d907d32a30cdbd97072c7d7c3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.39.1
loguru==0.7.2
lxml==5.3.0
lyft-dataset-sdk==0.0.8
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcls==0.25.0
mmcv==2.0.1
mmengine==0.10.4
mmsegmentation==0.30.0
model-index==0.1.11
mpmath==1.3.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
ninja==1.11.1.1
numba==0.56.4
numpy==1.23.5
nuscenes-devkit==1.1.11
oauthlib==3.2.2
omegaconf==2.3.0
omnidata-tools==0.0.23
onnx==1.16.1
onnxruntime==1.18.0
open3d==0.18.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
opt_einsum==3.4.0
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
plyfile==1.0.3
portalocker==2.10.1
prettytable==3.11.0
prompt_toolkit==3.0.47
protobuf==4.25.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.3
pydub==0.25.1
pyfiglet==0.8.post1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
pyperclip==1.9.0
pyquaternion==0.9.9
pyre-extensions==0.0.23
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==8.3.3
python-box==7.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytoolconfig==1.3.1
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
rope==1.13.0
rpds-py==0.19.0
rsa==4.9
ruff==0.9.6
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.13.2
semantic-version==2.10.0
Shapely==1.8.5.post1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.44.0
supervision==0.25.1
sympy==1.12.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.7
tokenizers==0.20.1
tomli==2.0.1
tomlkit==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.2
traitlets==5.14.3
transformers==4.45.2
trimesh==2.35.39
typer==0.15.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.33.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xformers==0.0.16
yacs==0.1.8
yapf==0.40.2
zipp==3.19.2
